requirements.txt
numpy
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.1
sklearn-pandas==2.2.0
xgboost